Update Tailwind CSS configuration and dependencies; refactor styles for improved readability and add fade-in animation
