+++ messages/fr.json
@@ -41,6 +41,7 @@
         "citation": "« La seule façon de faire du bon travail est d'aimer ce que l'on fait. » — Steve Jobs",
         "description": "Étudiant en informatique au sain de Bordeaux Ynov Campus. Je suis passionné par le développement web fullstack et l'exploration de diverses technologies. Je crée des solutions innovantes et performantes qui allient créativité, efficacité et maîtrise technique. J'ai eu l'occasion de porter plusieurs projets en tout genre, allant de la simple page web, à des applications plus complexes, en passant par le serveur de jeu.",
         "dates": "Naissance",
+        "born": "Je suis né le {date} à Paris. J'ai vécu 3 années à Paris avant de déménager à Bordeaux où je vis actuellement.",
         "contact": "Contact",
         "phone": "Téléphone",
         "bluesky": "Bluesky",
