+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@next/third-parties": "^15.1.6",
+    "@tailwindcss/postcss": "^4.0.3",
     "axios": "^1.7.9",
     "discord.js": "^14.17.2",
     "framer-motion": "^12.0.8",
@@ -29,8 +30,8 @@
     "eslint": "^9.19.0",
     "eslint-config-next": "15.1.6",
     "node-loader": "^2.1.0",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
+    "postcss": "^8.5.1",
+    "tailwindcss": "^4.0.3",
     "typescript": "5.7.3",
     "zlib-sync": "^0.1.9"
   },
