+++ src/app/[locale]/contact/page.tsx
@@ -8,7 +8,7 @@ export default function ContactPage() {
     return (
         <div className='flex flex-col min-h-screen bg-no-repeat bg-fixed overflow-x-hidden' style={{ backgroundImage: 'radial-gradient(circle, rgba(30, 172, 10, 0.7) 20%, rgba(0, 0, 0, 1) 80%)', backgroundSize: 'cover', backgroundPosition: 'top' }}>
             <div className='flex-grow mb-auto flex flex-col bg-fixed'>
-                <div className='h-screen w-full bg-opacity-60 bg-black flex flex-col justify-center items-center'>
+                <div className='h-screen w-full bg-black/80 flex flex-col justify-center items-center'>
                     <h1 className='text-white text-4xl md:text-6xl lg:text-8xl font-bold flex-grow flex items-center justify-center'>{t('title')}</h1>
                     <ScrollButton />
                 </div>
