+++ src/app/globals.css
@@ -1,6 +1,5 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+/* Import tailwind */
+@import "tailwindcss";
 
 .grecaptcha-badge {
     visibility: hidden !important;
@@ -27,4 +26,16 @@
 
 .nice-shadow {
     box-shadow: 0px 0px 50px 3px #ffffff80;
+}
+
+@keyframes fade-in {
+    from {
+        opacity: 0;
+    }
+    to {
+        opacity: 1;
+    }
+}
+.animate-fade-in {
+    animation: fade-in 1s ease-in-out;
 }
